Profit odds, Magic art, two assorted UPCs, parked Collection, compact hero
New 'What are the odds you profit?' section on every product page: a seeded
20,000-trial Monte Carlo over the product's own pull tables and card prices
(same sources/blend/graded options as the EV, so they can never disagree).
Shows P(beat market), P(beat MSRP), the median opening, and the top-decile
threshold, with the assumptions stated: uniform-within-tier reads as a
ceiling, selling is free, prices are today's. Engine lives in @packroi/ev
(profit.ts, 7 tests); fractional pack counts model assorted products.

Charizard UPC (16 SWSH-series packs, publisher-stated) and Terapagos ex UPC
(18 packs, 'may vary') are now ranked as uniform-draw blends across their
era's main sets -- the loader now allows fractional component counts for
exactly this. Both priced from eBay solds, art from Scrydex.

Magic sealed tiles drop the monochrome keyrune silhouette for the set's top
chase card art (Scryfall scans we already store) -- the best honest image
that exists for Magic sealed, which has no product-photo source.

Collection is parked until accounts exist: nav link removed, /collection
redirects home, page + APIs stay in git history.

Also: hero one notch smaller, tile caption reads 'average opening is ',
explicit {" "} where Next 16's JSX transform ate an inter-element space.

+++ apps/web/src/lib/jobs/refresh-catalog.ts
@@ -299,7 +299,7 @@ const productFileSchema = z.object({
        * setCode must be a ranked set (its own pull table drives the blend).
        */
       componentPacks: z
-        .array(z.object({ setCode: z.string().min(1), count: z.number().int().positive() }))
+        .array(z.object({ setCode: z.string().min(1), count: z.number().positive() }))
         .default([]),
       contentsNote: z.string().optional(),
     }),
